fix(sm): allow string for probes in check composition
The SM Check expects probe IDs for the probes, in Terraform the
`grafana_synthetic_monitoring_probes` data source can be used to grab
those based on the probe name. Crossplane (Upjet) doesn't have a 'data
source' concept so that is not possible.

The grafana-data function[1] aims to provide the translation from name
to ID. To accommodate for that, the XRD will need to allow for probes to
be names, then the function can replace those with IDs in the composite.

[1] https://github.com/grafana/crossplane-function-grafana-data/